Criação do teste para o parser e atualização da visao
