+++ controlador/parser/teste.cpp
@@ -1,0 +1,14 @@
+#include <iostream>
+#include <string>
+#include "parser.h"
+
+int main () {
+    std::string teste = "3+2+x+x*5";
+    expre::expre *arvore;
+
+    arvore = parser::parser(teste);
+
+    std::cout << arvore->calcular(10) << '\n';
+
+    return 0;
+}
